Made index.js less cluttered : - Created a helpers folder that has helping functions to index.js. - helpers folder exports corsOptions, serveSwaggerUI,isOriginAllowed functions
